+++ packages/frontend/src/components/council-topic/index.tsx
@@ -1,5 +1,5 @@
 'use client'
-import React, { useEffect, useState, useMemo, type FC } from 'react'
+import React, { useMemo, type FC } from 'react'
 // @twreporter
 import { TabletAndBelow } from '@twreporter/react-components/lib/rwd'
 import { CITY_LABEL } from '@twreporter/congress-dashboard-shared/lib/constants/city'
@@ -11,6 +11,7 @@ import Statistics from '@/components/council-topic/statistics'
 import TopicRelatedArticles from '@/components/topic/topic-related-articles'
 import FeedbackBlock from '@/components/layout/feedback-block'
 import ContentPageLayout from '../layout/content-page-without-filter-layout'
+import BackToTopButton from '@/components/councilor/back-to-top-button'
 // styles
 import {
   Spacing,
@@ -21,28 +22,25 @@ import {
 //  types
 import type { CouncilTopicFromRes } from '@/types/council-topic'
 import type { CouncilMeeting } from '@/types/council-meeting'
+import type { CouncilorWithWorkCounts } from '@/types/councilor'
 // custom hooks
 import useTopicData from '@/components/council-topic/hook/use-topic-data'
 
 type TopicPageProps = {
   topicData: CouncilTopicFromRes
+  councilorsData: CouncilorWithWorkCounts[]
   councilMeeting: CouncilMeeting
 }
 
 const CouncilTopicPage: FC<TopicPageProps> = ({
   topicData,
+  councilorsData,
   councilMeeting,
 }) => {
-  const [isLoading, setIsLoading] = useState<boolean>(false)
+  const { topic, speechCouncilorCount, billCouncilorCount, councilors } =
+    useTopicData(topicData, councilorsData)
 
-  const { topic, councilorCount, councilors, billsByCouncilor } =
-    useTopicData(topicData)
-
-  const pageTitle = `#${topic?.title} 的相關議案`
-
-  useEffect(() => {
-    setIsLoading(false)
-  }, [councilorCount, councilors, billsByCouncilor])
+  const pageTitle = `#${topic.title} 的相關發言與議案`
 
   const councilMeetingText = useMemo(
     () => `${CITY_LABEL[councilMeeting.city]}議會 | 第${councilMeeting.term}屆`,
@@ -56,16 +54,17 @@ const CouncilTopicPage: FC<TopicPageProps> = ({
           <TopicListContainer>
             <TopicList
               districtSlug={councilMeeting.city}
-              isLoading={isLoading}
+              councilMeetingId={Number(councilMeeting.id)}
               councilors={councilors}
-              billsByTopic={billsByCouncilor}
               topic={topic}
             />
           </TopicListContainer>
         </DesktopList>
         <DesktopAside>
           <Statistics
-            councilorCount={councilorCount}
+            speechCouncilorCount={speechCouncilorCount}
+            speechCount={topic.speechCount}
+            billCouncilorCount={billCouncilorCount}
             billCount={topic.billCount}
           />
           <TopicRelatedArticles
@@ -83,16 +82,17 @@ const CouncilTopicPage: FC<TopicPageProps> = ({
         </DesktopAside>
         <TabletAndBelow>
           <Statistics
-            councilorCount={councilorCount}
+            speechCouncilorCount={speechCouncilorCount}
+            speechCount={topic.speechCount}
+            billCouncilorCount={billCouncilorCount}
             billCount={topic.billCount}
           />
           <Spacing $height={32} />
           <TopicListContainer>
             <TopicList
               districtSlug={councilMeeting.city}
-              isLoading={isLoading}
+              councilMeetingId={Number(councilMeeting.id)}
               councilors={councilors}
-              billsByTopic={billsByCouncilor}
               topic={topic}
             />
           </TopicListContainer>
@@ -114,6 +114,7 @@ const CouncilTopicPage: FC<TopicPageProps> = ({
           <FeedbackBlock eventName="council-topic" />
         </TabletAndBelow>
       </ContentPageLayout>
+      <BackToTopButton />
     </>
   )
 }